Use xxHash32 primitives for HashBytes and extract shared dn_xxhash.h
Rewrite HashBytes to use xxHash32 QueueRound/MixFinal primitives instead
of byte-at-a-time DJB2. Extract xxHash32 code from typehashingalgorithms.h
into a shared inc/dn_xxhash.h header.
